refactor(spotify): extract SpotifyPagedQueue base class
SpotifyPlaylistQueue and SpotifyLikedSongsQueue were ~90% identical: the same
pagination state, fast-start windowing, parallel batched resolution,
source-level shuffle, hasNextPage/nextPage and fetch-next-page logic, differing
only in the fetch call and track extraction.

Hoist all shared logic into an abstract SpotifyPagedQueue; the two queues become
thin subclasses that supply fetchPage(offset, limit) and an optional
pre-provided list (providedTracks). Behavior-preserving — the fetch offset still
advances by raw page item count (PageResult.rawCount), not the filtered size, so
pagination is unchanged. Removes ~180 lines of duplicated logic; a pagination or
fast-start fix now lives in one place.
