+++ app/src/main/kotlin/com/metrolist/music/playback/queues/SpotifyLikedSongsQueue.kt
@@ -5,30 +5,20 @@
 
 package com.metrolist.music.playback.queues
 
-import androidx.media3.common.MediaItem
 import com.metrolist.music.models.MediaMetadata
 import com.metrolist.music.playback.SpotifyYouTubeMapper
 import com.metrolist.spotify.Spotify
 import com.metrolist.spotify.models.SpotifyTrack
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.withContext
-import timber.log.Timber
 
 /**
- * Queue implementation for Spotify Liked Songs (saved tracks).
- *
- * Optimized for fast playback start: only the selected track is resolved
- * during [getInitialStatus], while the remaining tracks are resolved
- * progressively in [nextPage] batches as the player approaches the end of
- * the currently loaded queue.
+ * Queue implementation for Spotify Liked Songs (saved tracks). All the
+ * pagination / fast-start / resolution logic lives in [SpotifyPagedQueue];
+ * this class only supplies the liked-songs fetch and the optional pre-provided list.
  */
 class SpotifyLikedSongsQueue(
-    private val startIndex: Int = 0,
-    private val mapper: SpotifyYouTubeMapper,
-    override val preloadItem: MediaMetadata? = null,
+    startIndex: Int = 0,
+    mapper: SpotifyYouTubeMapper,
+    preloadItem: MediaMetadata? = null,
     /**
      * Pre-ordered list of tracks to play, matching exactly what the UI displays (after
      * the user's sort/reverse). When provided, [startIndex] indexes into THIS list and no
@@ -36,207 +26,18 @@ class SpotifyLikedSongsQueue(
      * When null, the queue falls back to fetching from the Spotify API in its native order.
      */
     private val tracks: List<SpotifyTrack>? = null,
-) : Queue {
+) : SpotifyPagedQueue(startIndex, mapper, preloadItem) {
 
-    companion object {
-        private const val SPOTIFY_PAGE_SIZE = 50
-        private const val RESOLVE_BATCH_SIZE = 20
-        /** Resolve only the target + a few neighbors for instant playback start. */
-        private const val FAST_START_BEFORE = 0
-        private const val FAST_START_AFTER = 2
-    }
+    override val logTag: String = "SpotifyLikedSongsQueue"
 
-    // All Spotify tracks fetched so far (may span multiple API pages)
-    private val allTracks = mutableListOf<SpotifyTrack>()
+    override val providedTracks: List<SpotifyTrack>? = tracks?.filter { !it.isLocal }
 
-    // Index into [allTracks] for the next batch to resolve to YouTube
-    private var resolveOffset = 0
-
-    // Spotify API pagination state
-    private var apiFetchOffset = 0
-    private var apiTotal = 0
-    private var apiHasMore = true
-
-    override suspend fun getInitialStatus(): Queue.Status = withContext(Dispatchers.IO) {
-        try {
-            val provided = tracks
-            if (provided != null) {
-                // Play the exact list (and order) the UI is showing.
-                allTracks.clear()
-                allTracks.addAll(provided.filter { !it.isLocal })
-                apiTotal = allTracks.size
-                apiFetchOffset = allTracks.size
-                apiHasMore = false
-            } else {
-                val result = Spotify.likedSongs(limit = SPOTIFY_PAGE_SIZE, offset = 0).getOrThrow()
-                apiTotal = result.total
-                val fetched = result.items.map { it.track }.filter { !it.isLocal }
-                allTracks.addAll(fetched)
-                apiFetchOffset = result.items.size
-                apiHasMore = apiFetchOffset < apiTotal
-
-                while (startIndex >= allTracks.size && apiHasMore) {
-                    fetchNextApiPage()
-                }
-            }
-
-            if (allTracks.isEmpty()) {
-                return@withContext Queue.Status(title = null, items = emptyList(), mediaItemIndex = 0)
-            }
-
-            val targetIndex = startIndex.coerceIn(0, (allTracks.size - 1).coerceAtLeast(0))
-
-            // Fast-start: resolve only a tiny window (target + 2 next) for instant playback.
-            // The rest of the queue is populated via nextPage() in the background.
-            val windowStart = (targetIndex - FAST_START_BEFORE).coerceAtLeast(0)
-            val windowEnd = (targetIndex + FAST_START_AFTER + 1).coerceAtMost(allTracks.size)
-            val windowTracks = allTracks.subList(windowStart, windowEnd)
-
-            val resolvedItems = coroutineScope {
-                windowTracks.map { track -> async { mapper.resolveToMediaItem(track) } }
-                    .awaitAll()
-                    .filterNotNull()
-            }
-
-            if (resolvedItems.isEmpty()) {
-                Timber.w("SpotifyLikedSongsQueue: Could not resolve any track in initial window")
-                return@withContext Queue.Status(title = null, items = emptyList(), mediaItemIndex = 0)
-            }
-
-            resolveOffset = windowEnd
-
-            val mediaItemIndex = (targetIndex - windowStart)
-                .coerceIn(0, (resolvedItems.size - 1).coerceAtLeast(0))
-
-            Timber.d("SpotifyLikedSongsQueue: Fast-start resolved ${resolvedItems.size} tracks " +
-                "(window $windowStart..$windowEnd, target=$targetIndex, total=$apiTotal)")
-
-            Queue.Status(
-                title = null,
-                items = resolvedItems,
-                mediaItemIndex = mediaItemIndex,
-            )
-        } catch (e: Exception) {
-            Timber.e(e, "SpotifyLikedSongsQueue: Failed initial fetch")
-            Queue.Status(title = null, items = emptyList(), mediaItemIndex = 0)
-        }
-    }
-
-    override suspend fun getFullStatus(): Queue.Status? = withContext(Dispatchers.IO) {
-        try {
-            val provided = tracks
-            if (provided != null) {
-                // Use the exact UI-provided order; no pagination needed.
-                allTracks.clear()
-                allTracks.addAll(provided.filter { !it.isLocal })
-                apiTotal = allTracks.size
-                apiFetchOffset = allTracks.size
-                apiHasMore = false
-            } else {
-                // Build full list from scratch so we always include tracks 0..N (not just from startIndex onwards)
-                allTracks.clear()
-                apiFetchOffset = 0
-                apiHasMore = true
-                while (apiFetchOffset == 0 || apiHasMore) {
-                    if (apiFetchOffset == 0) {
-                        val result = Spotify.likedSongs(limit = SPOTIFY_PAGE_SIZE, offset = 0).getOrThrow()
-                        apiTotal = result.total
-                        val fetched = result.items.map { it.track }.filter { !it.isLocal }
-                        allTracks.addAll(fetched)
-                        apiFetchOffset = result.items.size
-                        apiHasMore = apiFetchOffset < apiTotal
-                    } else {
-                        fetchNextApiPage()
-                    }
-                }
-            }
-            if (allTracks.isEmpty()) return@withContext null
-            val targetIndex = startIndex.coerceIn(0, allTracks.size - 1)
-            // Resolve all tracks in parallel batches (order preserved) instead of
-            // one-by-one; on a large library the serial loop blocked shuffle-all for
-            // tens of seconds. Batch size matches nextPage() to bound concurrency.
-            val resolved = ArrayList<MediaItem?>(allTracks.size)
-            for (chunk in allTracks.chunked(RESOLVE_BATCH_SIZE)) {
-                resolved += coroutineScope {
-                    chunk.map { track -> async { mapper.resolveToMediaItem(track) } }.awaitAll()
-                }
-            }
-            val resolvedItems = mutableListOf<MediaItem>()
-            var mediaItemIndex = 0
-            for (i in allTracks.indices) {
-                if (i == targetIndex) mediaItemIndex = resolvedItems.size
-                resolved[i]?.let { resolvedItems.add(it) }
-            }
-            if (resolvedItems.isEmpty()) return@withContext null
-            mediaItemIndex = mediaItemIndex.coerceIn(0, resolvedItems.size - 1)
-            resolveOffset = allTracks.size
-            apiHasMore = false
-            Timber.d("SpotifyLikedSongsQueue: getFullStatus resolved ${resolvedItems.size} tracks (startIndex=$targetIndex)")
-            Queue.Status(title = null, items = resolvedItems, mediaItemIndex = mediaItemIndex)
-        } catch (e: Exception) {
-            Timber.e(e, "SpotifyLikedSongsQueue: getFullStatus failed")
-            null
-        }
-    }
-
-    override suspend fun shuffleRemainingTracks() = withContext(Dispatchers.IO) {
-        while (apiHasMore) {
-            fetchNextApiPage()
-        }
-        if (resolveOffset < allTracks.size) {
-            val remaining = allTracks.subList(resolveOffset, allTracks.size)
-            val shuffled = remaining.shuffled()
-            for (i in shuffled.indices) {
-                remaining[i] = shuffled[i]
-            }
-            Timber.d("SpotifyLikedSongsQueue: Shuffled ${remaining.size} remaining tracks " +
-                "(resolveOffset=$resolveOffset, total=${allTracks.size})")
-        }
-    }
-
-    override fun hasNextPage(): Boolean =
-        resolveOffset < allTracks.size || apiHasMore
-
-    override suspend fun nextPage(): List<MediaItem> = withContext(Dispatchers.IO) {
-        // If we've resolved all fetched tracks but the API has more, fetch another page
-        if (resolveOffset >= allTracks.size && apiHasMore) {
-            fetchNextApiPage()
-        }
-
-        if (resolveOffset >= allTracks.size) {
-            return@withContext emptyList()
-        }
-
-        // Resolve the next batch
-        val end = (resolveOffset + RESOLVE_BATCH_SIZE).coerceAtMost(allTracks.size)
-        val batch = allTracks.subList(resolveOffset, end)
-        resolveOffset = end
-
-        Timber.d("SpotifyLikedSongsQueue: Resolving batch of ${batch.size} tracks " +
-            "(offset=$resolveOffset/${allTracks.size}, apiTotal=$apiTotal)")
-
-        coroutineScope {
-            batch.map { track -> async { mapper.resolveToMediaItem(track) } }
-                .awaitAll()
-                .filterNotNull()
-        }
-    }
-
-    private suspend fun fetchNextApiPage() {
-        if (!apiHasMore) return
-        try {
-            val result = Spotify.likedSongs(
-                limit = SPOTIFY_PAGE_SIZE,
-                offset = apiFetchOffset,
-            ).getOrThrow()
-            val fetched = result.items.map { it.track }.filter { !it.isLocal }
-            allTracks.addAll(fetched)
-            apiFetchOffset += result.items.size
-            apiHasMore = apiFetchOffset < apiTotal
-            Timber.d("SpotifyLikedSongsQueue: Fetched API page, now have ${allTracks.size} tracks")
-        } catch (e: Exception) {
-            Timber.e(e, "SpotifyLikedSongsQueue: Failed to fetch next API page")
-            apiHasMore = false
-        }
+    override suspend fun fetchPage(offset: Int, limit: Int): PageResult {
+        val result = Spotify.likedSongs(limit = limit, offset = offset).getOrThrow()
+        return PageResult(
+            tracks = result.items.map { it.track }.filter { !it.isLocal },
+            total = result.total,
+            rawCount = result.items.size,
+        )
     }
 }
